おみくじを導入するために電卓機能を独立させた
